requirements.txt
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
